Fix type/descriptor mismatch in CDR tests
Union discriminat in type declared as 32-bit, in descriptor as 8-bit.  The union cases
require alignment to 32-bit and so this works fine on a little-endian machine and only
breaks on a big-endian one.

+++ src/core/ddsc/tests/cdrstream.c
@@ -634,7 +634,7 @@ static void sample_free_opt (void *s)
 /* @appendable */
 typedef struct TestIdl_AppendableUnion0
 {
-  int32_t _d;
+  int8_t _d;
   union
   {
     uint32_t field1;
